+++ src/Subscribers/ShopgateDetectTrait.php
@@ -2,6 +2,7 @@
 
 namespace Shopgate\WebcheckoutSW6\Subscribers;
 
+use Shopgate\WebcheckoutSW6\SgateWebcheckoutSW6;
 use Symfony\Component\HttpFoundation\Request;
 
 trait ShopgateDetectTrait
@@ -27,5 +28,12 @@ trait ShopgateDetectTrait
         $sgSession = $hasSession && $request->getSession()->get(IsShopgateSubscriber::SG_SESSION_KEY, 0);
 
         return $sgAgent || $sgSession || $sgCookie;
+    }
+
+    public function isShopgateApiCall(Request $request): bool
+    {
+        return $request->headers->has(SgateWebcheckoutSW6::IS_SHOPGATE_CHECK) &&
+            $request->headers->has('sw-context-token') &&
+            $request->headers->has('sw-access-key');
     }
 }